+++ lib/stoplight/evaluation_strategy.rb
@@ -3,7 +3,7 @@
 module Stoplight
   class EvaluationStrategy
     # @param config [Stoplight::Light::Config]
-    # @param metadata [Stoplight::DataStore::Metadata]
+    # @param metadata [Stoplight::Metadata]
     # @return [Boolean]
     def evaluate(config, metadata)
       [metadata.consecutive_failures, metadata.failures].min >= config.threshold
